:truck: move create preview assessment uscase
