+++ api/src/shared/application/assessments/assessment-controller.js
@@ -30,11 +30,6 @@ const save = async function (request, h, dependencies = { assessmentRepository }
   return h.response(assessmentSerializer.serialize(createdAssessment)).created();
 };
 
-const createAssessmentPreviewForPix1d = async function (request, h, dependencies = { assessmentSerializer }) {
-  const createdAssessment = await usecases.createPreviewAssessment({});
-  return h.response(dependencies.assessmentSerializer.serialize(createdAssessment)).created();
-};
-
 const get = async function (request, _, dependencies = { assessmentSerializer }) {
   const assessmentId = request.params.id;
   const locale = extractLocaleFromRequest(request);
@@ -166,7 +161,6 @@ const assessmentController = {
   updateLastChallengeState,
   findCompetenceEvaluations,
   autoValidateNextChallenge,
-  createAssessmentPreviewForPix1d,
   createCertificationChallengeLiveAlert,
 };
