+++ api/src/shared/application/assessments/index.js
@@ -32,15 +32,6 @@ const register = async function (server) {
       },
     },
     {
-      method: 'POST',
-      path: '/api/pix1d/assessments/preview',
-      config: {
-        auth: false,
-        handler: assessmentController.createAssessmentPreviewForPix1d,
-        tags: ['api', 'pix1d', 'assessment'],
-      },
-    },
-    {
       method: 'GET',
       path: '/api/assessments/{id}/next',
       config: {
