+++ api/tests/school/integration/domain/usecases/create-preview-assessment_test.js
@@ -1,7 +1,7 @@
-import { createPreviewAssessment } from '../../../../lib/domain/usecases/create-preview-assessment.js';
-import { Assessment } from '../../../../src/shared/domain/models/index.js';
-import * as assessmentRepository from '../../../../src/shared/infrastructure/repositories/assessment-repository.js';
-import { expect, knex } from '../../../test-helper.js';
+import { createPreviewAssessment } from '../../../../../src/school/domain/usecases/create-preview-assessment.js';
+import { Assessment } from '../../../../../src/shared/domain/models/index.js';
+import * as assessmentRepository from '../../../../../src/shared/infrastructure/repositories/assessment-repository.js';
+import { expect, knex } from '../../../../test-helper.js';
 
 describe('Integration | UseCases | create-preview-assessment', function () {
   let assessmentId;
