+++ api/tests/school/unit/application/assessment-route_test.js
@@ -83,4 +83,21 @@ describe('Unit | Application | Router | assessment-router', function () {
       expect(getCurrentActivityStub).to.have.been.called;
     });
   });
+
+  describe('POST /api/pix1d/assessments/preview', function () {
+    it('should return 200', async function () {
+      // given
+      sinon
+        .stub(assessmentController, 'createAssessmentPreviewForPix1d')
+        .callsFake((request, h) => h.response('ok').code(200));
+      const httpTestServer = new HttpTestServer();
+      await httpTestServer.register(moduleUnderTest);
+
+      // when
+      const response = await httpTestServer.request('POST', '/api/pix1d/assessments/preview');
+
+      // then
+      expect(response.statusCode).to.equal(200);
+    });
+  });
 });
