+++ api/tests/shared/unit/application/assessements/assessment-controller_test.js
@@ -11,22 +11,6 @@ import { featureToggles } from '../../../../../src/shared/infrastructure/feature
 import { domainBuilder, expect, hFake, sinon } from '../../../../test-helper.js';
 
 describe('Unit | Controller | assessment-controller', function () {
-  describe('#createAssessmentPreviewForPix1d', function () {
-    it('should call the expected usecase', async function () {
-      const assessmentSerializer = { serialize: sinon.stub() };
-      const createdAssessment = Symbol('created-assessment');
-      assessmentSerializer.serialize.withArgs(createdAssessment).resolves(Symbol('serialized-assessment'));
-      sinon.stub(usecases, 'createPreviewAssessment').resolves(createdAssessment);
-
-      const result = await assessmentController.createAssessmentPreviewForPix1d({}, hFake, {
-        assessmentSerializer,
-      });
-
-      expect(result.statusCode).to.be.equal(201);
-      expect(assessmentSerializer.serialize).to.have.been.calledWithExactly(createdAssessment);
-    });
-  });
-
   describe('#get', function () {
     const authenticatedUserId = '12';
     const locale = 'fr';
