+++ api/tests/shared/unit/application/assessements/index_test.js
@@ -46,23 +46,6 @@ describe('Unit | Application | Router | assessment-router', function () {
     });
   });
 
-  describe('POST /api/pix1d/assessments/preview', function () {
-    it('should return 200', async function () {
-      // given
-      sinon
-        .stub(assessmentController, 'createAssessmentPreviewForPix1d')
-        .callsFake((request, h) => h.response('ok').code(200));
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const response = await httpTestServer.request('POST', '/api/pix1d/assessments/preview');
-
-      // then
-      expect(response.statusCode).to.equal(200);
-    });
-  });
-
   describe('GET /api/assessments/{id}/next', function () {
     it('should return 200', async function () {
       // given
